Concrete implementations

SimpleGraphs

GraphsBase.SimpleGraphs.SimpleDiGraph — Type

mutable struct SimpleDiGraph{T<:Integer} <: AbstractGraph{T<:Integer, GraphsBase.SimpleGraphs.SimpleEdge{T<:Integer}}

A type representing a directed graph with unweighted, non-multiple edges.

Fields

ne::Int64: Number of edges

fadjlist::Array{Vector{T}, 1} where T<:Integer: Forward adjacency list such that F[v] contains the out-neighbors of vertex v

badjlist::Array{Vector{T}, 1} where T<:Integer: Backward adjacency list such that B[u] contains the in-neighbors of vertex v

GraphsBase.SimpleGraphs.SimpleEdge — Type

struct SimpleEdge{T<:Integer} <: AbstractEdge{T<:Integer, Int64}

A type representing an unweighted directed edge.

Fields

src::Integer: Source of the edge

dst::Integer: Destination of the edge

GraphsBase.SimpleGraphs.SimpleGraph — Type

mutable struct SimpleGraph{T<:Integer} <: AbstractGraph{T<:Integer, GraphsBase.SimpleGraphs.SimpleEdge{T<:Integer}}

A type representing an undirected graph with unweighted, non-multiple edges.

Fields

ne::Int64: Number of edges

adjlist::Array{Vector{T}, 1} where T<:Integer: Adjacency list such that L[v] contains the neighbors of vertex v

SimpleWeightedGraphs

GraphsBase.SimpleWeightedGraphs.SimpleWeightedDiGraph — Type

struct SimpleWeightedDiGraph{T<:Integer, W} <: AbstractGraph{T<:Integer, GraphsBase.SimpleWeightedGraphs.SimpleWeightedEdge{T<:Integer, W}}

A type representing a directed graph with weighted, non-multiple edges.

Fields

weights::SparseArrays.SparseMatrixCSC{W, T} where {T<:Integer, W}: Transposed weighted adjacency matrix: weights[v, u] contains the weight of edge (u, v)

GraphsBase.SimpleWeightedGraphs.SimpleWeightedEdge — Type

struct SimpleWeightedEdge{T<:Integer, W} <: AbstractEdge{T<:Integer, W}

A type representing a weighted directed edge.

Fields

src::Integer: Source of the edge

dst::Integer: Destination of the edge

weight::Any: Weight of the edge

GraphsBase.SimpleWeightedGraphs.SimpleWeightedGraph — Type

struct SimpleWeightedGraph{T<:Integer, W} <: AbstractGraph{T<:Integer, GraphsBase.SimpleWeightedGraphs.SimpleWeightedEdge{T<:Integer, W}}

A type representing an undirected graph with weighted, non-multiple edges.

Fields

weights::SparseArrays.SparseMatrixCSC{W, T} where {T<:Integer, W}: Symmetric weighted adjacency matrix: weights[u, v] and weights[v, u] both contain the weight of edge {u, v}